Add CLI tool to deal with JWT

+++ src/authn/retrieve.rs
@@ -127,8 +127,11 @@ pub fn call(
             }
         })
         .and_then(|(expires_in, identity, refresh_token, namespace)| {
-            let payload =
-                jwt::AccessToken::new(namespace.label.clone(), expires_in, identity.account_id);
+            let payload = jwt::AccessToken::new(
+                namespace.label.clone(),
+                u32::from(expires_in),
+                identity.account_id,
+            );
             let access_token = jwt::AccessToken::encode(payload)?;
 
             let payload = jwt::RefreshToken::new(namespace.label, identity.account_id);
